+++ jupyterhub/orm.py
@@ -4,6 +4,7 @@
 # Distributed under the terms of the Modified BSD License.
 
 from datetime import datetime
+import enum
 import json
 
 from tornado import gen
@@ -14,7 +15,7 @@ from sqlalchemy.types import TypeDecorator, TEXT
 from sqlalchemy import (
     inspect,
     Column, Integer, ForeignKey, Unicode, Boolean,
-    DateTime,
+    DateTime, Enum
 )
 from sqlalchemy.ext.declarative import declarative_base, declared_attr
 from sqlalchemy.orm import sessionmaker, relationship, backref
@@ -607,6 +608,53 @@ class APIToken(Base):
         db.add(orm_token)
         db.commit()
         return token
+
+
+#------------------------------------
+# OAuth tables
+#------------------------------------
+
+
+class GrantType(enum.Enum):
+    authorization_code = 'authorization_code'
+    implicit = 'implicit'
+    password = 'password'
+    client_credentials = 'client_credentials'
+    refresh_token = 'refresh_token'
+
+
+class OAuthAccessToken(Base):
+    __tablename__ = 'oauth_access_tokens'
+    id = Column(Integer, primary_key=True, autoincrement=True)
+
+    client_id = Column(Unicode(1023))
+    grant_type = Column(Enum(GrantType), nullable=False)
+    expires_at = Column(Integer)
+    refresh_token = Column(Unicode(36))
+    refresh_expires_at = Column(Integer)
+    user_id = Column(Integer, ForeignKey('users.id', ondelete='CASCADE'))
+    user = relationship(User)
+    api_token_id = Column(Integer, ForeignKey('api_tokens.id', ondelete='CASCADE'))
+    api_token = relationship(APIToken, backref='oauth_token')
+
+
+class OAuthCode(Base):
+    __tablename__ = 'oauth_codes'
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    client_id = Column(Unicode(1023))
+    code = Column(Unicode(36))
+    expires_at = Column(Integer)
+    redirect_uri = Column(Unicode(1023))
+    user_id = Column(Integer, ForeignKey('users.id', ondelete='CASCADE'))
+
+
+class OAuthClient(Base):
+    __tablename__ = 'oauth_clients'
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    identifier = Column(Unicode(1023), unique=True)
+    secret = Column(Unicode(1023))
+    redirect_uri = Column(Unicode(1023))
+
 
 
 def new_session_factory(url="sqlite:///:memory:", reset=False, **kwargs):
